refactor(catalog): hoist shared CLI test fixtures, drop dead imports
Following the test_cli.py split, `runner` and `catalog_with_source_and_transform`
were duplicated across files. Hoist both to conftest.py so there is a single
source of truth.

Also drop now-rotting `# Tests for: ...` header comments and the unused
imports the split left behind. The two `--pdb` tests in test_cli_run_compose.py
that constructed CliRunner() inline now use the shared `runner` fixture.

Measured per-file totals (single-process, sequential) for reference:
run_compose 67 s, inspect_git 38 s, init_crud 37 s, params 23 s. The split
is uneven; rebalancing run_compose vs. params is a possible follow-up.
