+++ python/xorq/catalog/tests/test_cli_params.py
@@ -1,6 +1,4 @@
-# Tests for: --rename-params, -p/--params, ExprBuilder entries, run-cached with params
 import pytest
-from click.testing import CliRunner
 
 import xorq.api as xo
 import xorq.expr.builders as _builders_mod
@@ -16,11 +14,6 @@ from xorq.vendor.ibis.expr import operations as ops
 
 
 @pytest.fixture
-def runner():
-    yield CliRunner()
-
-
-@pytest.fixture
 def saved_registry():
     """Save and restore the handler registry around a test."""
     saved = dict(_FROM_TAG_NODE_REGISTRY)
@@ -31,24 +24,6 @@ def saved_registry():
     _FROM_TAG_NODE_REGISTRY.update(saved)
     _builders_mod._BUILTIN_KEYS = saved_keys
     _builders_mod._initialized = saved_init
-
-
-@pytest.fixture
-def catalog_with_source_and_transform(catalog_path):
-    """Populate a catalog with a source entry and an unbound transform entry."""
-    catalog = Catalog.from_kwargs(path=catalog_path, init=False)
-
-    source = xo.memtable(
-        {"user_id": [1, 2, 3], "amount": [10.0, 20.0, 30.0], "name": ["a", "b", "c"]}
-    )
-    source_entry = catalog.add(source, aliases=("src",))
-
-    schema = source.schema()
-    unbound = ops.UnboundTable(name="placeholder", schema=schema).to_expr()
-    transform = unbound.filter(unbound.amount > 0).select("user_id", "amount")
-    transform_entry = catalog.add(transform, aliases=("trn",))
-
-    return catalog_path, source_entry.name, transform_entry.name
 
 
 @pytest.fixture
