+++ python/xorq/catalog/tests/test_cli_run_compose.py
@@ -1,39 +1,10 @@
-# Tests for: run, compose, compose+run roundtrip, --pdb flag, run-cached commands
 from pathlib import Path
 from unittest.mock import MagicMock
 
 import pyarrow as pa
-import pytest
-from click.testing import CliRunner
 
-import xorq.api as xo
-from xorq.catalog.catalog import Catalog
 from xorq.catalog.cli import cli
 from xorq.cli import cli as top_cli
-from xorq.vendor.ibis.expr import operations as ops
-
-
-@pytest.fixture
-def runner():
-    yield CliRunner()
-
-
-@pytest.fixture
-def catalog_with_source_and_transform(catalog_path):
-    """Populate a catalog with a source entry and an unbound transform entry."""
-    catalog = Catalog.from_kwargs(path=catalog_path, init=False)
-
-    source = xo.memtable(
-        {"user_id": [1, 2, 3], "amount": [10.0, 20.0, 30.0], "name": ["a", "b", "c"]}
-    )
-    source_entry = catalog.add(source, aliases=("src",))
-
-    schema = source.schema()
-    unbound = ops.UnboundTable(name="placeholder", schema=schema).to_expr()
-    transform = unbound.filter(unbound.amount > 0).select("user_id", "amount")
-    transform_entry = catalog.add(transform, aliases=("trn",))
-
-    return catalog_path, source_entry.name, transform_entry.name
 
 
 # --- run command ---
@@ -403,14 +374,13 @@ def test_compose_then_run_arrow_stdout(runner, catalog_with_source_and_transform
 # --- --pdb flag ---
 
 
-def test_pdb_flag_invokes_post_mortem(tmp_path, monkeypatch):
+def test_pdb_flag_invokes_post_mortem(runner, tmp_path, monkeypatch):
     """--pdb should let exceptions propagate to PdbGroup, which calls pdb.post_mortem."""
     mock_pm = MagicMock()
     monkeypatch.setattr("xorq.cli.pdb.post_mortem", mock_pm)
 
     # "catalog list" on a non-catalog directory fails inside the command body
     # (not at Click arg-parsing time), so it exercises click_context_catalog.
-    runner = CliRunner()
     result = runner.invoke(
         top_cli,
         ["--pdb", "catalog", "--path", str(tmp_path), "list"],
@@ -419,9 +389,8 @@ def test_pdb_flag_invokes_post_mortem(tmp_path, monkeypatch):
     assert mock_pm.called, "pdb.post_mortem was not called with --pdb"
 
 
-def test_no_pdb_flag_wraps_exception(tmp_path):
+def test_no_pdb_flag_wraps_exception(runner, tmp_path):
     """Without --pdb, errors should be wrapped as clean 'Error: ...' messages."""
-    runner = CliRunner()
     result = runner.invoke(
         top_cli,
         ["catalog", "--path", str(tmp_path), "list"],
